Rename to coldfront_plugin_cloud
