+++ Vagrantfile
@@ -1,5 +1,5 @@
 Vagrant.configure("2") do |config|
-    config.vm.synced_folder ".", "/home/vagrant/coldfront-plugin-openstack/"
+    config.vm.synced_folder ".", "/home/vagrant/coldfront-plugin-cloud/"
     config.vm.network :private_network
 
     config.vm.define "openstack" do |openstack|
@@ -14,7 +14,7 @@ Vagrant.configure("2") do |config|
         openstack.vm.provision "shell", privileged: false, inline: <<-SHELL
             set -xe
 
-            cd ~/coldfront-plugin-openstack
+            cd ~/coldfront-plugin-cloud
             ./ci/devstack.sh
             ./ci/setup.sh
             ./ci/run_functional_tests.sh
@@ -33,7 +33,7 @@ Vagrant.configure("2") do |config|
         openshift.vm.provision "shell", privileged: false, inline: <<-SHELL
             set -xe
 
-            cd ~/coldfront-plugin-openstack
+            cd ~/coldfront-plugin-cloud
             ./ci/microshift.sh
             ./ci/setup.sh
             ./ci/run_functional_tests_openshift.sh
