+++ ci/run_functional_tests_openshift.sh
@@ -14,4 +14,4 @@ export DJANGO_SETTINGS_MODULE="local_settings"
 export FUNCTIONAL_TESTS="True"
 export OS_AUTH_URL="https://onboarding-onboarding.cluster.local"
 
-coldfront test coldfront_plugin_openstack.tests.functional.openshift
+coldfront test coldfront_plugin_cloud.tests.functional.openshift
