+++ ci/run_functional_tests_openstack.sh
@@ -27,6 +27,6 @@ export KEYCLOAK_USER="admin"
 export KEYCLOAK_PASS="nomoresecret"
 export KEYCLOAK_REALM="master"
 
-coldfront test coldfront_plugin_openstack.tests.functional.openstack
+coldfront test coldfront_plugin_cloud.tests.functional.openstack
 
 openstack application credential delete $OPENSTACK_DEVSTACK_APPLICATION_CREDENTIAL_ID
