+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = coldfront_plugin_openstack
+name = coldfront_plugin_cloud
 version = 0.1.0
 author = Kristi Nikolla
 author_email = knikolla@bu.edu
