+++ src/coldfront_plugin_cloud/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'coldfront_plugin_cloud.apps.OpenStackConfig'
