+++ src/coldfront_plugin_cloud/apps.py
@@ -2,7 +2,7 @@ from django.apps import AppConfig
 
 
 class OpenStackConfig(AppConfig):
-    name = 'coldfront_plugin_openstack'
+    name = 'coldfront_plugin_cloud'
 
     def ready(self):
-        import coldfront_plugin_openstack.signals
+        import coldfront_plugin_cloud.signals
