+++ src/coldfront_plugin_cloud/attributes.py
