+++ src/coldfront_plugin_cloud/config.py
@@ -1,7 +1,7 @@
 from coldfront.config.base import INSTALLED_APPS
 from coldfront.config.env import ENV
 
-if 'coldfront_plugin_openstack' not in INSTALLED_APPS:
+if 'coldfront_plugin_cloud' not in INSTALLED_APPS:
     INSTALLED_APPS += [
-        'coldfront_plugin_openstack',
+        'coldfront_plugin_cloud',
     ]
