--- src/coldfront_plugin_openstack/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'coldfront_plugin_openstack.apps.OpenStackConfig'
