+++ src/local_settings.py
@@ -3,6 +3,6 @@ import pkgutil
 from coldfront.config.settings import *
 from django.conf import settings
 
-plugin_openstack = pkgutil.get_loader('coldfront_plugin_openstack.config')
+plugin_openstack = pkgutil.get_loader('coldfront_plugin_cloud.config')
 
 include(plugin_openstack.get_filename())
